+++ example/web_frontend/src/public/assets/js/model/template-scripts.js
@@ -1,0 +1,14 @@
+jQuery(document).ready(function($) {  
+  // Owl Carousel                     
+  var owl = $('.carousel-fade-transition');
+  owl.owlCarousel({
+    nav: true,
+    dots: true,
+    items: 1,
+    loop: true,
+    navText: ["&#xe605","&#xe606"],
+    autoplay: true, 
+    animateOut: 'fadeOut',
+    autoplayTimeout: 5000
+  });
+});